requirements.txt
# pigar doesn't work :/
django == 3.2.18
datetime == 5
pyyaml == 6.0
psycopg2 == 2.9.5
channels == 4.0.0
channels_redis == 4.0.0
django-debug-toolbar == 3.8.1
environs == 9.5.0
gunicorn == 20.1.0
djangorestframework-jwt == 1.11.0
djangorestframework == 3.14.0
uvicorn[standard] == 0.20.0
pytest
pytest-django
drf_spectacular
pytest-asyncio
drf-nested-routers
daphne
